python/requirements.txt
flask>=3.0
requests>=2.31
python-dotenv>=1.0
PyYAML>=6.0
